requirements.txt
# sap-ai-sdk-gen 은 구 generative-ai-hub-sdk 의 rebranded 배포명.
# 6.1.2+ 부터 gen_ai_hub.orchestration_v2 (config_ref 방식) 지원.
# main_v2.py / run_batch_v2.py 가 이 모듈을 요구.
sap-ai-sdk-gen>=6.1.2
python-dotenv>=1.0.0
httpx>=0.27
requests>=2.31.0
